fix(stream): widen sum_sq to i128 so variance/z-score can't silently corrupt on large amounts

+++ crates/features/src/derive.rs
@@ -88,7 +88,7 @@ mod tests {
         )
     }
 
-    fn aggregates_30d(count: u64, sum_minor: i64, sum_sq: i64) -> WindowAggregates {
+    fn aggregates_30d(count: u64, sum_minor: i64, sum_sq: i128) -> WindowAggregates {
         WindowAggregates {
             windows: vec![WindowStat {
                 label: "30d".to_string(),
